+++ VpnHood.Server.Access/Providers/FileAccessServerProvider/FileAccessServer.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Net;
+using System.Net.Sockets;
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
 using System.Text.Json;
@@ -62,6 +63,16 @@ public class FileAccessServer : IAccessServer
     {
         ServerInfo = serverInfo;
         ServerStatus = serverInfo.Status;
+
+        // update UdpEndPoints if they are not configured 
+        var udpEndPoints = ServerConfig.UdpEndPointsValue.ToArray();
+        foreach (var udpEndPoint in udpEndPoints.Where(x => x.Port == 0))
+        {
+            udpEndPoint.Port = udpEndPoint.AddressFamily == AddressFamily.InterNetworkV6
+                ? serverInfo.FreeUdpPortV6 : serverInfo.FreeUdpPortV4;
+        }
+        ServerConfig.UdpEndPoints = udpEndPoints.Where(x => x.Port != 0).ToArray();
+
         return Task.FromResult(ServerConfig);
     }
 
@@ -77,10 +88,19 @@ public class FileAccessServer : IAccessServer
         if (accessItem == null)
             return new SessionResponseEx(SessionErrorCode.AccessError) { ErrorMessage = "Token does not exist." };
 
-        return SessionManager.CreateSession(sessionRequestEx, accessItem);
+        var ret = SessionManager.CreateSession(sessionRequestEx, accessItem);
+
+        // set endpoints
+        ret.TcpEndPoints = new[] { sessionRequestEx.HostEndPoint };
+        ret.UdpEndPoints = ServerConfig.UdpEndPointsValue
+            .Where(x => x.AddressFamily == sessionRequestEx.HostEndPoint.AddressFamily)
+            .Select(x => new IPEndPoint(sessionRequestEx.HostEndPoint.Address, x.Port))
+            .ToArray();
+
+        return ret;
     }
 
-    public async Task<SessionResponseEx> Session_Get(uint sessionId, IPEndPoint hostEndPoint, IPAddress? clientIp)
+    public async Task<SessionResponseEx> Session_Get(ulong sessionId, IPEndPoint hostEndPoint, IPAddress? clientIp)
     {
         _ = hostEndPoint;
         _ = clientIp;
@@ -107,17 +127,17 @@ public class FileAccessServer : IAccessServer
         return SessionManager.GetSession(sessionId, accessItem, hostEndPoint);
     }
 
-    public Task<SessionResponseBase> Session_AddUsage(uint sessionId, Traffic traffic)
+    public Task<SessionResponseBase> Session_AddUsage(ulong sessionId, Traffic traffic)
     {
         return Session_AddUsage(sessionId, traffic, false);
     }
 
-    public Task<SessionResponseBase> Session_Close(uint sessionId, Traffic traffic)
+    public Task<SessionResponseBase> Session_Close(ulong sessionId, Traffic traffic)
     {
         return Session_AddUsage(sessionId, traffic, true);
     }
 
-    private async Task<SessionResponseBase> Session_AddUsage(uint sessionId, Traffic traffic, bool closeSession)
+    private async Task<SessionResponseBase> Session_AddUsage(ulong sessionId, Traffic traffic, bool closeSession)
     {
         // find token
         var tokenId = SessionManager.TokenIdFromSessionId(sessionId);