+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v2.9.369
+### Client & Server
+* Feature: Use shared UDP port 
+* Feature: Improve protocol anonymity and anti-fingerprinting for UDP 
+* Update: Improve security, performance, and battery usage
+* Update: Remove excessive error logs on disconnecting
+
 # v2.8.361
 ### Server
 * Feature: Enable hot reconfig for VpnHood Server TCP listener to avoid unnecessary restarts on unchanged endpoints